Relongify and replace %d with %ld for longs

+++ vex/vexf_put.c
@@ -8,7 +8,7 @@
 #ifdef F2C
 #include "f2c.h"
 #else
-typedef int integer;
+typedef long int integer;
 typedef int ftnlen;
 typedef double doublereal;
 #endif
